libsnapshot: Add a way to detach snapuserd.
Currently, when init unlinks /dev/socket/snapuserd, the daemon currently
polling the socket will fail in accept4(). This works because the
process won't exit until all threads have terminated, but it's
confusing, and it creates log spam.

Instead, add an explicit "detach" message so that snapuserd will stop
listening for new connections and will immediately join on remaining
threads. This way snapuserd will gracefully exit when all threads are
complete, and we're guaranteed no new threads will be created.

This will only be used for first-stage instances of snapuserd.

Bug: 173476209
Test: full OTA with VABC, no accept4() errors
Change-Id: Ibeffa4a35043ed43a70166eeee86204e2b3d03aa